DL-2683 Updated Dependencies
