+++ conf/application.conf
@@ -65,7 +65,7 @@ application.langs="en"
 # you may need to define a router file `conf/my.application.routes`.
 # Default to Routes in the root package (and conf/routes)
 # !!!WARNING!!! DO NOT CHANGE THIS ROUTER
-application.router=prod.Routes
+play.http.router=prod.Routes
 
 
 # Controller
